+++ docs/src/content/en/docs/deployment/_meta.ts
@@ -1,6 +1,7 @@
 const meta = {
   overview: "Overview",
   "server-deployment": "Server deployment",
+  monorepo: "With a Monorepo",
   "web-framework": "With a Web Framework",
   "serverless-platforms": "Serverless Platforms",
   "cloud-providers": "Cloud Providers",