+++ packages/core/src/vector/index.ts
@@ -1,2 +1,3 @@
 export * from './vector';
 export * from './types';
+export * from './embed';